Split panels into smart/dumb components for better separation of concerns
