+++ src/features/report_details/ReportFightDetails.tsx
@@ -1,8 +1,6 @@
-import OpenInNewIcon from '@mui/icons-material/OpenInNew';
-import { Paper, Typography, Button, CircularProgress, Box, Stack, Tooltip } from '@mui/material';
 import React from 'react';
 import { useSelector } from 'react-redux';
-import { useNavigate, useSearchParams } from 'react-router-dom';
+import { useSearchParams } from 'react-router-dom';
 
 import { useAuth } from '../../AuthContext';
 import { useReportFightParams } from '../../hooks/useReportFightParams';
@@ -12,13 +10,12 @@ import { fetchReportMasterData, clearMasterData } from '../../store/master_data/
 import { fetchReportData } from '../../store/report/reportSlice';
 import { useAppDispatch } from '../../store/useAppDispatch';
 
-import FightDetails from './FightDetails';
+import ReportFightDetailsView from './ReportFightDetailsView';
 
 const ReportFightDetails: React.FC = () => {
   const { reportId, fightId } = useReportFightParams();
   const [searchParams] = useSearchParams();
   const dispatch = useAppDispatch();
-  const navigate = useNavigate();
 
   const { accessToken } = useAuth();
 
@@ -62,85 +59,24 @@ const ReportFightDetails: React.FC = () => {
     }
   }, [fight, reportId, accessToken, dispatch]);
 
-  // Show loading panel if fights, master data, or events are loading, or fights are missing
-  if (
-    fightsLoading ||
-    fights.length === 0 ||
-    masterDataLoading ||
-    !masterDataLoaded ||
-    (eventsLoading && currentFetchFightId === Number(fightId))
-  ) {
-    return (
-      <Box
-        sx={{
-          display: 'flex',
-          flexDirection: 'column',
-          alignItems: 'center',
-          justifyContent: 'center',
-          minHeight: 200,
-        }}
-      >
-        <CircularProgress sx={{ mb: 2 }} />
-        <Typography variant="h6">Loading...</Typography>
-      </Box>
-    );
-  }
-  if (fightsError) {
-    return <Typography color="error">Error loading fights: {fightsError}</Typography>;
-  }
-  if (!fight) {
-    return <Typography variant="h6">Fight not found.</Typography>;
-  }
-
   // Get selectedTabId from query param if present
   const selectedTabId = searchParams.has('selectedTabId')
     ? Number(searchParams.get('selectedTabId'))
     : undefined;
 
   return (
-    <Paper elevation={2} sx={{ p: 3 }}>
-      <Button
-        variant="outlined"
-        sx={{ mb: 2 }}
-        onClick={() => {
-          navigate(`/report/${reportId}`);
-        }}
-      >
-        Back to Fight List
-      </Button>
-
-      <Stack direction="row" alignItems="center" justifyContent="space-between" sx={{ mb: 1 }}>
-        <Typography variant="h6" gutterBottom={false}>
-          Fight Details
-        </Typography>
-
-        <Tooltip title="View full report on ESO Logs">
-          <Button
-            component="a"
-            href={`https://www.esologs.com/reports/${reportId}`}
-            target="_blank"
-            rel="noopener noreferrer"
-            variant="outlined"
-            size="small"
-            startIcon={<OpenInNewIcon />}
-            sx={{
-              textTransform: 'none',
-              fontSize: '0.875rem',
-            }}
-          >
-            ESO Logs
-          </Button>
-        </Tooltip>
-      </Stack>
-
-      {fight && (
-        <Typography variant="body2" color="text.secondary" sx={{ mb: 2 }}>
-          Total Fight Time: {((fight.endTime - fight.startTime) / 1000).toFixed(1)} seconds
-        </Typography>
-      )}
-
-      <FightDetails fight={fight} selectedTabId={selectedTabId} />
-    </Paper>
+    <ReportFightDetailsView
+      fight={fight}
+      fightsLoading={fightsLoading || fights.length === 0}
+      fightsError={fightsError}
+      masterDataLoading={masterDataLoading}
+      masterDataLoaded={masterDataLoaded}
+      eventsLoading={eventsLoading}
+      currentFetchFightId={currentFetchFightId ?? null}
+      selectedTabId={selectedTabId}
+      reportId={reportId}
+      fightId={fightId}
+    />
   );
 };
