+++ src/features/report_details/critical_damage/CriticalDamagePanel.tsx
@@ -1,4 +1,3 @@
-import { Box } from '@mui/material';
 import React from 'react';
 import { useSelector } from 'react-redux';
 
@@ -6,50 +5,49 @@ import { FightFragment } from '../../../graphql/generated';
 import { RootState } from '../../../store/storeWithHistory';
 import { resolveActorName } from '../../../utils/resolveActorName';
 
-import PlayerCriticalDamageDetails from './PlayerCriticalDamageDetails';
+import CriticalDamagePanelView from './CriticalDamagePanelView';
 
 interface CriticalDamagePanelProps {
   fight: FightFragment;
 }
 
+/**
+ * Smart component that handles data processing and state management for critical damage panel
+ */
 const CriticalDamagePanel: React.FC<CriticalDamagePanelProps> = ({ fight }) => {
   // Get report actors from masterData
   const actorsById = useSelector((state: RootState) => state.masterData.actorsById);
 
-  // Filter for Player actors only
-  const playerActors = Object.values(actorsById).filter((actor) => actor.type === 'Player');
+  // Process player data
+  const players = React.useMemo(() => {
+    return Object.values(actorsById)
+      .filter((actor) => actor.type === 'Player' && actor.id)
+      .map((actor) => ({
+        id: String(actor.id),
+        name: String(resolveActorName(actor)),
+      }));
+  }, [actorsById]);
 
   // Track which panels are expanded
   const [expandedPanels, setExpandedPanels] = React.useState<Record<string, boolean>>({});
 
-  const handleExpandChange =
+  const handleExpandChange = React.useCallback(
     (playerId: string) => (event: React.SyntheticEvent, isExpanded: boolean) => {
       setExpandedPanels((prev) => ({
         ...prev,
         [playerId]: isExpanded,
       }));
-    };
+    },
+    []
+  );
 
   return (
-    <Box>
-      {playerActors.map((actor) => {
-        if (!actor.id) return null;
-
-        const playerId = String(actor.id);
-        const playerName = String(resolveActorName(actor));
-
-        return (
-          <PlayerCriticalDamageDetails
-            key={playerId}
-            id={playerId}
-            name={playerName}
-            fight={fight}
-            expanded={expandedPanels[playerId] || false}
-            onExpandChange={handleExpandChange(playerId)}
-          />
-        );
-      })}
-    </Box>
+    <CriticalDamagePanelView
+      players={players}
+      fight={fight}
+      expandedPanels={expandedPanels}
+      onExpandChange={handleExpandChange}
+    />
   );
 };
